nfc_model/backend/requirements.txt
flask
flask-pymongo
pymongo
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
pillow
requests
